paddlespatial/networks/vmrgae/requirements.txt
numpy~=1.21.4
ipaddlepaddle-gpu==2.1.3
pgl==2.1.5